+++ demos/bookstore/app/auth.tsx
@@ -91,7 +91,7 @@ export default {
         let headers = new Headers()
         setSessionCookie(headers, session.sessionId)
 
-        return redirect(routes.account.index, { headers })
+        return redirect(routes.account.index.href(), { headers })
       },
     },
 
@@ -173,7 +173,7 @@ export default {
         let headers = new Headers()
         setSessionCookie(headers, session.sessionId)
 
-        return redirect(routes.account.index, { headers })
+        return redirect(routes.account.index.href(), { headers })
       },
     },
 
@@ -181,7 +181,7 @@ export default {
       let session = getSession(request)
       logout(session.sessionId)
 
-      return redirect(routes.home)
+      return redirect(routes.home.href())
     },
 
     forgotPassword: {